mythic-cli refactor
