+++ Mythic_CLI/src/cmd/config/env.go
@@ -1,18 +1,101 @@
-package internal
+package config
 
 import (
+	"bufio"
 	"fmt"
+	"github.com/MythicMeta/Mythic_CLI/cmd/utils"
+	"github.com/spf13/viper"
 	"log"
 	"os"
 	"path/filepath"
 	"sort"
 	"strings"
-
-	"github.com/spf13/viper"
 )
 
+var MythicPossibleServices = []string{
+	"mythic_postgres",
+	"mythic_react",
+	"mythic_server",
+	"mythic_nginx",
+	"mythic_rabbitmq",
+	"mythic_graphql",
+	"mythic_documentation",
+	"mythic_jupyter",
+	"mythic_sync",
+	"mythic_grafana",
+	"mythic_prometheus",
+	"mythic_postgres_exporter",
+}
 var mythicEnv = viper.New()
 
+// GetIntendedMythicServiceNames uses MythicEnv host values for various services to see if they should be local or remote
+func GetIntendedMythicServiceNames() ([]string, error) {
+	// need to see about adding services back in if they were for remote hosts before
+	containerList := []string{}
+	for _, service := range MythicPossibleServices {
+		// service is a mythic service, but it's not in our current container list (i.e. not in docker-compose)
+		switch service {
+		case "mythic_react":
+			if mythicEnv.GetString("MYTHIC_REACT_HOST") == "127.0.0.1" || mythicEnv.GetString("MYTHIC_REACT_HOST") == "mythic_react" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_nginx":
+			if mythicEnv.GetString("NGINX_HOST") == "127.0.0.1" || mythicEnv.GetString("NGINX_HOST") == "mythic_nginx" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_rabbitmq":
+			if mythicEnv.GetString("RABBITMQ_HOST") == "127.0.0.1" || mythicEnv.GetString("RABBITMQ_HOST") == "mythic_rabbitmq" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_server":
+			if mythicEnv.GetString("MYTHIC_SERVER_HOST") == "127.0.0.1" || mythicEnv.GetString("MYTHIC_SERVER_HOST") == "mythic_server" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_postgres":
+			if mythicEnv.GetString("POSTGRES_HOST") == "127.0.0.1" || mythicEnv.GetString("POSTGRES_HOST") == "mythic_postgres" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_graphql":
+			if mythicEnv.GetString("HASURA_HOST") == "127.0.0.1" || mythicEnv.GetString("HASURA_HOST") == "mythic_graphql" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_documentation":
+			if mythicEnv.GetString("DOCUMENTATION_HOST") == "127.0.0.1" || mythicEnv.GetString("DOCUMENTATION_HOST") == "mythic_documentation" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_jupyter":
+			if mythicEnv.GetString("JUPYTER_HOST") == "127.0.0.1" || mythicEnv.GetString("JUPYTER_HOST") == "mythic_jupyter" {
+				containerList = append(containerList, service)
+			}
+		case "mythic_grafana":
+			if mythicEnv.GetBool("postgres_debug") {
+				containerList = append(containerList, service)
+			}
+		case "mythic_prometheus":
+			if mythicEnv.GetBool("postgres_debug") {
+				containerList = append(containerList, service)
+			}
+		case "mythic_postgres_exporter":
+			if mythicEnv.GetBool("postgres_debug") {
+				containerList = append(containerList, service)
+			}
+			/*
+				case "mythic_sync":
+					if mythicSyncPath, err := filepath.Abs(filepath.Join(utils.GetCwdFromExe(), InstalledServicesFolder, "mythic_sync")); err != nil {
+						fmt.Printf("[-] Failed to get the absolute path to mythic_sync: %v\n", err)
+					} else if _, err = os.Stat(mythicSyncPath); !os.IsNotExist(err) {
+						// this means that the mythic_sync folder _does_ exist
+						containerList = append(containerList, service)
+					}
+
+			*/
+		}
+	}
+	return containerList, nil
+}
+func GetMythicEnv() *viper.Viper {
+	return mythicEnv
+}
 func setMythicConfigDefaultValues() {
 	// global configuration
 	mythicEnv.SetDefault("debug_level", "warning")
@@ -54,7 +137,7 @@ func setMythicConfigDefaultValues() {
 	mythicEnv.SetDefault("postgres_bind_localhost_only", true)
 	mythicEnv.SetDefault("postgres_db", "mythic_db")
 	mythicEnv.SetDefault("postgres_user", "mythic_user")
-	mythicEnv.SetDefault("postgres_password", generateRandomPassword(30))
+	mythicEnv.SetDefault("postgres_password", utils.GenerateRandomPassword(30))
 	mythicEnv.SetDefault("postgres_cpus", "2")
 	mythicEnv.SetDefault("postgres_mem_limit", "")
 	mythicEnv.SetDefault("postgres_bind_local_mount", true)
@@ -63,18 +146,18 @@ func setMythicConfigDefaultValues() {
 	mythicEnv.SetDefault("rabbitmq_port", 5672)
 	mythicEnv.SetDefault("rabbitmq_bind_localhost_only", true)
 	mythicEnv.SetDefault("rabbitmq_user", "mythic_user")
-	mythicEnv.SetDefault("rabbitmq_password", generateRandomPassword(30))
+	mythicEnv.SetDefault("rabbitmq_password", utils.GenerateRandomPassword(30))
 	mythicEnv.SetDefault("rabbitmq_vhost", "mythic_vhost")
 	mythicEnv.SetDefault("rabbitmq_cpus", "2")
 	mythicEnv.SetDefault("rabbitmq_mem_limit", "")
 	mythicEnv.SetDefault("rabbitmq_bind_local_mount", true)
 	// jwt configuration
-	mythicEnv.SetDefault("jwt_secret", generateRandomPassword(30))
+	mythicEnv.SetDefault("jwt_secret", utils.GenerateRandomPassword(30))
 	// hasura configuration
 	mythicEnv.SetDefault("hasura_host", "mythic_graphql")
 	mythicEnv.SetDefault("hasura_port", 8080)
 	mythicEnv.SetDefault("hasura_bind_localhost_only", true)
-	mythicEnv.SetDefault("hasura_secret", generateRandomPassword(30))
+	mythicEnv.SetDefault("hasura_secret", utils.GenerateRandomPassword(30))
 	mythicEnv.SetDefault("hasura_cpus", "2")
 	mythicEnv.SetDefault("hasura_mem_limit", "2gb")
 	mythicEnv.SetDefault("hasura_bind_local_mount", true)
@@ -83,7 +166,7 @@ func setMythicConfigDefaultValues() {
 	mythicEnv.SetDefault("REBUILD_ON_START", true)
 	// Mythic instance configuration
 	mythicEnv.SetDefault("mythic_admin_user", "mythic_admin")
-	mythicEnv.SetDefault("mythic_admin_password", generateRandomPassword(30))
+	mythicEnv.SetDefault("mythic_admin_password", utils.GenerateRandomPassword(30))
 	mythicEnv.SetDefault("default_operation_name", "Operation Chimera")
 	mythicEnv.SetDefault("allowed_ip_blocks", "0.0.0.0/0,::/0")
 	mythicEnv.SetDefault("default_operation_webhook_url", "")
@@ -114,10 +197,10 @@ func parseMythicEnvironmentVariables() {
 	setMythicConfigDefaultValues()
 	mythicEnv.SetConfigName(".env")
 	mythicEnv.SetConfigType("env")
-	mythicEnv.AddConfigPath(getCwdFromExe())
+	mythicEnv.AddConfigPath(utils.GetCwdFromExe())
 	mythicEnv.AutomaticEnv()
-	if !fileExists(filepath.Join(getCwdFromExe(), ".env")) {
-		_, err := os.Create(filepath.Join(getCwdFromExe(), ".env"))
+	if !utils.FileExists(filepath.Join(utils.GetCwdFromExe(), ".env")) {
+		_, err := os.Create(filepath.Join(utils.GetCwdFromExe(), ".env"))
 		if err != nil {
 			log.Fatalf("[-] .env doesn't exist and couldn't be created\n")
 		}
@@ -178,7 +261,7 @@ func writeMythicEnvironmentVariables() {
 		keys = append(keys, k)
 	}
 	sort.Strings(keys)
-	f, err := os.Create(filepath.Join(getCwdFromExe(), ".env"))
+	f, err := os.Create(filepath.Join(utils.GetCwdFromExe(), ".env"))
 	if err != nil {
 		log.Fatalf("[-] Error writing out environment!\n%v", err)
 	}
@@ -233,8 +316,72 @@ func SetConfigStrings(key string, value string) {
 	mythicEnv.Get(key)
 	writeMythicEnvironmentVariables()
 }
+func GetBuildArguments() []string {
+	var buildEnv = viper.New()
+	buildEnv.SetConfigName("build.env")
+	buildEnv.SetConfigType("env")
+	buildEnv.AddConfigPath(utils.GetCwdFromExe())
+	buildEnv.AutomaticEnv()
+	if !utils.FileExists(filepath.Join(utils.GetCwdFromExe(), "build.env")) {
+		log.Printf("[*] No build.env file detected in Mythic's root directory; not supplying build arguments to docker containers\n")
+		log.Printf("    If you need to supply build arguments to docker containers, create build.env and supply key=value entries there\n")
+		return []string{}
+	}
+	if err := buildEnv.ReadInConfig(); err != nil {
+		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
+			log.Fatalf("[-] Error while reading in build.env file: %s\n", err)
+		} else {
+			log.Fatalf("[-]Error while parsing build.env file: %s\n", err)
+		}
+	}
+	c := buildEnv.AllSettings()
+	// to make it easier to read and look at, get all the keys, sort them, and display variables in order
+	keys := make([]string, 0, len(c))
+	for k := range c {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+	var args []string
+	for _, key := range keys {
+		args = append(args, fmt.Sprintf("%s=%s", strings.ToUpper(key), buildEnv.GetString(key)))
+	}
+	return args
+}
+
+// https://gist.github.com/r0l1/3dcbb0c8f6cfe9c66ab8008f55f8f28b
+func AskConfirm(prompt string) bool {
+	reader := bufio.NewReader(os.Stdin)
+	for {
+		fmt.Printf("%s [y/n]: ", prompt)
+		input, err := reader.ReadString('\n')
+		if err != nil {
+			fmt.Printf("[-] Failed to read user input\n")
+			return false
+		}
+		input = strings.ToLower(strings.TrimSpace(input))
+		if input == "y" || input == "yes" {
+			return true
+		} else if input == "n" || input == "no" {
+			return false
+		}
+	}
+}
+
+// https://gist.github.com/r0l1/3dcbb0c8f6cfe9c66ab8008f55f8f28b
+func AskVariable(prompt string, environmentVariable string) {
+	reader := bufio.NewReader(os.Stdin)
+	for {
+		fmt.Printf("%s: ", prompt)
+		input, err := reader.ReadString('\n')
+		if err != nil {
+			fmt.Printf("[-] Failed to read user input\n")
+		}
+		input = strings.TrimSpace(input)
+		mythicEnv.Set(environmentVariable, input)
+		writeMythicEnvironmentVariables()
+	}
+}
 func Initialize() {
 	parseMythicEnvironmentVariables()
 	writeMythicEnvironmentVariables()
-	CheckDockerCompose()
 }
