+++ Mythic_CLI/src/cmd/internal/reset.go
@@ -9,7 +9,7 @@ import (
 
 func DatabaseReset() {
 	fmt.Printf("[*] Stopping Mythic\n")
-	DockerStop([]string{})
+	ServiceStop([]string{})
 	workingPath := getCwdFromExe()
 	fmt.Printf("[*] Removing database files\n")
 	if mythicEnv.GetBool("postgres_bind_local_mount") {
@@ -20,7 +20,7 @@ func DatabaseReset() {
 			fmt.Printf("[+] Successfully reset datbase files\n")
 		}
 	} else {
-		DockerRemoveContainers([]string{"mythic_postgres"})
+		ServiceRemoveContainers([]string{"mythic_postgres"})
 		err := DockerRemoveVolume("mythic_postgres_volume")
 		if err != nil {
 			fmt.Printf("[-] Failed to remove database:\n%v\n", err)
@@ -31,7 +31,7 @@ func DatabaseReset() {
 func RabbitmqReset(explicitCall bool) {
 	if explicitCall {
 		fmt.Printf("[*] Stopping Mythic\n")
-		DockerStop([]string{})
+		ServiceStop([]string{})
 		fmt.Printf("[*] Removing rabbitmq files\n")
 	}
 	if mythicEnv.GetBool("rabbitmq_bind_local_mount") {
@@ -45,7 +45,7 @@ func RabbitmqReset(explicitCall bool) {
 			}
 		}
 	} else {
-		DockerRemoveContainers([]string{"mythic_rabbitmq"})
+		ServiceRemoveContainers([]string{"mythic_rabbitmq"})
 		err := DockerRemoveVolume("mythic_rabbitmq_volume")
 		if err != nil {
 			fmt.Printf("[-] Failed to remove rabbitmq files:\n%v\n", err)
