+++ mythic-docker/Dockerfile
@@ -1,6 +1,6 @@
 #FROM ghcr.io/its-a-feature/mythic_server:v0.0.2.32
 
-FROM ghcr.io/its-a-feature/mythic_server:v0.0.2.33
+FROM itsafeaturemythic/mythic_go_base:latest
 
 WORKDIR /usr/src/app
 
@@ -14,7 +14,7 @@ COPY ["src/", "."]
 
 RUN make build_final
 
-FROM ghcr.io/its-a-feature/mythic_server:v0.0.2.33
+FROM alpine
 
 COPY --from=0 /mythic_server /mythic_server
